move trivia utils functions to a separate file
