+++ commands/music/music-trivia.js
@@ -1,5 +1,11 @@
 const { SlashCommandBuilder } = require('@discordjs/builders');
 const { MessageEmbed } = require('discord.js');
+const {
+  getRandom,
+  normalizeValue,
+  capitalizeWords,
+  getLeaderBoard
+} = require('../../utils/trivia/utilFunctions');
 const fs = require('fs');
 
 module.exports = {
@@ -35,7 +41,7 @@ module.exports = {
       );
     }
 
-    const jsonSongs = fs.readFileSync('././utils/music/songs.json', 'utf-8');
+    const jsonSongs = fs.readFileSync('././utils/trivia/songs.json', 'utf-8');
 
     const songsArray = getRandom(JSON.parse(jsonSongs), 5);
 
@@ -73,23 +79,6 @@ module.exports = {
     playTrivia(interaction.channel, player, songsArray, score);
   }
 };
-
-function getRandom(arr, n) {
-  var result = new Array(n),
-    len = arr.length,
-    taken = new Array(len);
-  if (n > len)
-    throw new RangeError('getRandom: more elements taken than available!');
-  while (n--) {
-    var x = Math.floor(Math.random() * len);
-    // prettier-ignore
-    result[n] = arr[(x in taken) ? taken[x] : x];
-    // prettier-ignore
-    taken[x] = (--len in taken) ? taken[len] : len;
-    // prettier-ignore-end
-  }
-  return result;
-}
 
 async function playTrivia(textChannel, player, songsArray, score) {
   // Randomize a number but one that won't be too close to the track ending
@@ -201,9 +190,9 @@ async function playTrivia(textChannel, player, songsArray, score) {
       })
     );
 
-    const song = `${capitalize_Words(
+    const song = `${capitalizeWords(
       songsArray[0].singers[0]
-    )}: ${capitalize_Words(songsArray[0].title)}`;
+    )}: ${capitalizeWords(songsArray[0].title)}`;
 
     const embed = new MessageEmbed()
       .setColor('#ff7373')
@@ -231,34 +220,3 @@ async function playTrivia(textChannel, player, songsArray, score) {
     return playTrivia(textChannel, player, songsArray, score);
   });
 }
-
-var normalizeValue = value =>
-  value
-    .normalize('NFD')
-    .replace(/[\u0300-\u036f]/g, '') // remove diacritics
-    .replace(/[^0-9a-zA-Z\s]/g, '') // remove non-alphanumeric characters
-    .trim()
-    .replace(/\s+/g, ' ')
-    .toLowerCase(); // remove duplicate spaces
-
-var capitalize_Words = str => {
-  return str.replace(/\w\S*/g, function (txt) {
-    return txt.charAt(0).toUpperCase() + txt.substr(1).toLowerCase();
-  });
-};
-
-var getLeaderBoard = arr => {
-  if (!arr) return;
-  if (!arr[0]) return; // issue #422
-  let leaderBoard = '';
-
-  leaderBoard = `👑   **${arr[0][0]}:** ${arr[0][1]}  points`;
-
-  if (arr.length > 1) {
-    for (let i = 1; i < arr.length; i++) {
-      leaderBoard =
-        leaderBoard + `\n\n   ${i + 1}: ${arr[i][0]}: ${arr[i][1]}  points`;
-    }
-  }
-  return leaderBoard;
-};
